GeoJSON writer split into feature files (#5508)
* Rename ogr.thematic.structure to writer.thematic.structure
* Sonar findings
* Split GeoJSON files into thematic structure
* Create thematic structure unit test
* Documentation
* Copyright
* Updated MultiFileWriter unit test, removed 'Geometry' and 'FCODE' replaced with 'Thematic'
* Update unit tests, services, and translations

+++ hoot-core/src/main/cpp/hoot/core/conflate/tile/TileUtils.cpp
@@ -51,7 +51,7 @@ int TileUtils::getRandomTileIndex(const std::vector<geos::geom::Envelope>& tiles
 
   const size_t numBboxes = tiles.size();
   LOG_VARD(numBboxes);
-  return Tgs::Random::instance()->generateInt(numBboxes);
+  return Tgs::Random::instance()->generateInt(static_cast<int>(numBboxes));
 }
 
 geos::geom::Envelope TileUtils::getRandomTile(const std::vector<geos::geom::Envelope>& tiles, int randomSeed)
